requirements.txt
fpdf2==2.7.8
pytest==8.0.0
tabulate==0.9.0
